+++ NEWS.md
@@ -1,5 +1,6 @@
 # lmeInfo 0.1.3.9999
 
+* Added an option to return separate level-1 variance components for models that use `weights = varIdent(form = ~ 1 | Stratum)`.
 * Generalized the `g_mlm()` function to allow use of separate models for the numerator and denominator of the effect size.
 * Modified the stored results of `g_mlm()` so that the `returnModel` argument is no longer necessary.